refactor: extract holochain management functionality into standalone crate 'holochain-manager', modify tauri-plugin-holochain to use holochain-manager under-the-hood
