+++ crates/hc-pilot/src/main.rs
@@ -9,7 +9,8 @@ use lair_keystore::dependencies::sodoken::{BufRead, BufWrite};
 use std::collections::HashMap;
 use std::path::PathBuf;
 use tauri::{AppHandle, Context, Wry};
-use tauri_plugin_holochain::{HolochainExt, HolochainPluginConfig, WANNetworkConfig};
+use tauri_plugin_holochain::HolochainExt;
+use holochain_manager::{HolochainManagerConfig, WANNetworkConfig};
 use url2::url2;
 
 #[derive(Parser, Debug)]
@@ -105,7 +106,7 @@ fn main() {
         )
         .plugin(tauri_plugin_holochain::init(
             vec_to_locked(password.as_bytes().to_vec()).expect("Can't build passphrase"),
-            HolochainPluginConfig {
+            HolochainManagerConfig {
                 wan_network_config,
                 holochain_dir: conductor_dir,
             },
@@ -157,7 +158,7 @@ async fn setup(
     membrane_proofs: HashMap<String, std::sync::Arc<holochain_types::prelude::SerializedBytes>>,
     network_seed: Option<String>,
 ) -> anyhow::Result<AppInfo> {
-    let admin_ws = handle.holochain()?.admin_websocket().await?;
+    let admin_ws = handle.holochain()?.holochain_runtime.admin_websocket().await?;
     let agent_key = match agent_key {
         Some(agent_key) => agent_key,
         None => {
