+++ crates/holochain-manager/src/commands/mod.rs
@@ -1,0 +1,8 @@
+mod install_web_app;
+pub use install_web_app::{UpdateAppError, update_app, install_app, install_web_app};
+
+mod list_apps;
+pub use list_apps::list_apps;
+
+mod sign_zome_call;
+pub use sign_zome_call::{sign_zome_call, sign_zome_call_with_client, ZomeCallUnsignedTauri};
