+++ crates/holochain-manager/src/launch/mdns.rs
@@ -19,7 +19,8 @@ pub async fn spawn_mdns_bootstrap(admin_port: u16) -> crate::Result<()> {
                 "Could not connect to websocket: {err:?}"
             ))
         })?;
-    tauri::async_runtime::spawn(async move {
+
+    tokio::spawn(async move {
         let mut spaces_listened_to: HashSet<KitsuneSpace> = HashSet::new();
         let mut cells_ids_broadcasted: HashMap<
             (KitsuneSpace, KitsuneAgent),
@@ -29,13 +30,13 @@ pub async fn spawn_mdns_bootstrap(admin_port: u16) -> crate::Result<()> {
             let Ok(agent_infos) = admin_ws.agent_info(None).await else {
                 continue;
             };
-
+    
             // let cell_info: Vec<CellInfo> =
             let spaces: HashSet<KitsuneSpace> = agent_infos
                 .iter()
                 .map(|agent_info| agent_info.space.as_ref().clone())
                 .collect();
-
+    
             for space in spaces {
                 if !spaces_listened_to.contains(&space) {
                     if let Err(err) = spawn_listen_to_space_task(space.clone(), admin_port).await {
@@ -45,7 +46,7 @@ pub async fn spawn_mdns_bootstrap(admin_port: u16) -> crate::Result<()> {
                     spaces_listened_to.insert(space);
                 }
             }
-
+    
             for agent_info in agent_infos {
                 let cell_id = (
                     agent_info.space.as_ref().clone(),
@@ -71,7 +72,7 @@ pub async fn spawn_mdns_bootstrap(admin_port: u16) -> crate::Result<()> {
                 // store handle in self
                 cells_ids_broadcasted.insert(cell_id, handle);
             }
-
+    
             async_std::task::sleep(Duration::from_secs(5)).await;
         }
     });
@@ -89,52 +90,50 @@ pub async fn spawn_listen_to_space_task(space: KitsuneSpace, admin_port: u16) ->
         })?;
     let space_b64 = base64::prelude::BASE64_URL_SAFE_NO_PAD.encode(&space[..]);
 
-    tauri::async_runtime::spawn(async move {
-        let stream = mdns_listen(space_b64);
-        tokio::pin!(stream);
-        while let Some(maybe_response) = stream.next().await {
-            match maybe_response {
-                Ok(response) => {
-                    log::info!("Peer found via MDNS {:?}", response);
-                    // Decode response
-                    let maybe_agent_info_signed: Result<AgentInfoSigned, std::io::Error> =
-                        rmp_decode(&mut &*response.buffer);
-                    if let Err(e) = maybe_agent_info_signed {
-                        log::error!("Failed to decode MDNS peer {:?}", e);
+    let stream = mdns_listen(space_b64);
+    tokio::pin!(stream);
+    while let Some(maybe_response) = stream.next().await {
+        match maybe_response {
+            Ok(response) => {
+                log::info!("Peer found via MDNS {:?}", response);
+                // Decode response
+                let maybe_agent_info_signed: Result<AgentInfoSigned, std::io::Error> =
+                    rmp_decode(&mut &*response.buffer);
+                if let Err(e) = maybe_agent_info_signed {
+                    log::error!("Failed to decode MDNS peer {:?}", e);
+                    continue;
+                }
+                if let Ok(remote_agent_info_signed) = maybe_agent_info_signed {
+                    // Add to local storage
+                    let Ok(agent_infos) = admin_ws.agent_info(None).await else {
                         continue;
-                    }
-                    if let Ok(remote_agent_info_signed) = maybe_agent_info_signed {
-                        // Add to local storage
-                        let Ok(agent_infos) = admin_ws.agent_info(None).await else {
-                            continue;
-                        };
+                    };
 
-                        if agent_infos
-                            .iter()
-                            .find(|agent_info| {
-                                remote_agent_info_signed
-                                    .agent
-                                    .as_ref()
-                                    .eq(agent_info.agent.as_ref())
-                            })
-                            .is_none()
+                    if agent_infos
+                        .iter()
+                        .find(|agent_info| {
+                            remote_agent_info_signed
+                                .agent
+                                .as_ref()
+                                .eq(agent_info.agent.as_ref())
+                        })
+                        .is_none()
+                    {
+                        log::error!("Adding agent info {remote_agent_info_signed:?}");
+                        if let Err(e) = admin_ws
+                            .add_agent_info(vec![remote_agent_info_signed])
+                            .await
                         {
-                            log::error!("Adding agent info {remote_agent_info_signed:?}");
-                            if let Err(e) = admin_ws
-                                .add_agent_info(vec![remote_agent_info_signed])
-                                .await
-                            {
-                                log::error!("Failed to store MDNS peer {:?}", e);
-                            }
+                            log::error!("Failed to store MDNS peer {:?}", e);
                         }
                     }
                 }
-                Err(e) => {
-                    log::error!("Failed to get peers from MDNS {:?}", e);
-                }
+            }
+            Err(e) => {
+                log::error!("Failed to get peers from MDNS {:?}", e);
             }
         }
-    });
+    }
 
     Ok(())
 }
