+++ crates/holochain-manager/src/launch/signal.rs
