+++ crates/tauri-plugin-holochain/build.rs
@@ -8,7 +8,5 @@ const COMMANDS: &[&str] = &[
 
 fn main() {
     tauri_plugin::Builder::new(COMMANDS)
-        // .android_path("android")
-        // .ios_path("ios")
         .build();
 }
