+++ crates/tauri-plugin-holochain/src/commands/list_apps.rs
@@ -4,12 +4,5 @@ use tauri::{command, AppHandle, Runtime};
 
 #[command]
 pub(crate) async fn list_apps<R: Runtime>(app: AppHandle<R>) -> crate::Result<Vec<AppInfo>> {
-    let admin_ws = app.holochain()?.admin_websocket().await?;
-
-    let apps = admin_ws
-        .list_apps(None)
-        .await
-        .map_err(|err| crate::Error::ConductorApiError(err))?;
-
-    Ok(apps)
+    Ok(holochain_manager::commands::list_apps(app.holochain()?.holochain_runtime.clone()).await?)
 }
